changed setting to settings to make more sense
